+++ docs/api/admin/code_samples/JavaScript/admin_products_{id}_variants/post.js
@@ -1,8 +1,8 @@
 import Medusa from "@medusajs/medusa-js"
 const medusa = new Medusa({ baseUrl: MEDUSA_BACKEND_URL, maxRetries: 3 })
 // must be previously logged in or use api token
-medusa.admin.products.createVariant(product_id, {
-  title: 'Color',
+medusa.admin.products.createVariant(productId, {
+  title: "Color",
   prices: [
     {
       amount: 1000,
@@ -12,11 +12,11 @@ medusa.admin.products.createVariant(product_id, {
   options: [
     {
       option_id,
-      value: 'S'
+      value: "S"
     }
   ],
   inventory_quantity: 100
 })
 .then(({ product }) => {
   console.log(product.id);
-});
+})