+++ docs/api/store/code_samples/Shell/store_customers/post.sh
@@ -1,5 +1,5 @@
-curl --location --request POST 'https://medusa-url.com/store/customers' \
---header 'Content-Type: application/json' \
+curl -X POST 'https://medusa-url.com/store/customers' \
+-H 'Content-Type: application/json' \
 --data-raw '{
     "first_name": "Alec",
     "last_name": "Reynolds",